Add gallery transformations
