+++ src/blocks/gallery-masonry/index.js
@@ -2,7 +2,6 @@
  * External dependencies
  */
 import classnames from 'classnames';
-import filter from 'lodash/filter';
 
 /**
  * Internal dependencies
@@ -11,8 +10,9 @@ import './styles/style.scss';
 import './styles/editor.scss';
 import icons from './icons';
 import edit from './edit';
+import transforms from './transforms';
 import { BackgroundStyles } from '../../components/block-gallery/background/';
-import { GalleryAttributes, GalleryTransforms, GalleryClasses, GalleryStyles } from '../../components/block-gallery/shared';
+import { GalleryAttributes, GalleryClasses, GalleryStyles } from '../../components/block-gallery/shared';
 
 /**
  * WordPress dependencies
@@ -59,111 +59,19 @@ const settings = {
 		align: [ 'wide', 'full' ],
 	},
 
-	transforms: {
-		from: [
-			{
-				type: 'block',
-				blocks: [ 'blockgallery/stacked' ],
-				transform: ( attributes ) => (
-					createBlock( `coblocks/${ name }`, {
-						...GalleryTransforms( attributes ),
-					} )
-				),
-			},
-			{
-				type: 'block',
-				blocks: [ 'blockgallery/carousel' ],
-				transform: ( attributes ) => (
-					createBlock( `coblocks/${ name }`, {
-						...GalleryTransforms( attributes ),
-					} )
-				),
-			},
-			{
-				type: 'block',
-				blocks: [ 'blockgallery/thumbnails' ],
-				transform: ( attributes ) => (
-					createBlock( `coblocks/${ name }`, {
-						...GalleryTransforms( attributes ),
-					} )
-				),
-			},
-			{
-				type: 'block',
-				blocks: [ 'blockgallery/offset' ],
-				transform: ( attributes ) => (
-					createBlock( `coblocks/${ name }`, {
-						...GalleryTransforms( attributes ),
-					} )
-				),
-			},
-			{
-				type: 'block',
-				blocks: [ 'blockgallery/auto-height' ],
-				transform: ( attributes ) => (
-					createBlock( `coblocks/${ name }`, {
-						...GalleryTransforms( attributes ),
-					} )
-				),
-			},
-			{
-				type: 'block',
-				blocks: [ 'core/gallery' ],
-				transform: ( attributes ) => (
-					createBlock( `coblocks/${ name }`, {
-						...GalleryTransforms( attributes ),
-					} )
-				),
-			},
-			{
-				type: 'block',
-				isMultiBlock: true,
-				blocks: [ 'core/image' ],
-				transform: ( attributes ) => {
-					const validImages = filter( attributes, ( { id, url } ) => id && url );
-					if ( validImages.length > 0 ) {
-						return createBlock( `coblocks/${ name }`, {
-							images: validImages.map( ( { id, url, alt, caption } ) => ( { id, url, alt, caption } ) ),
-							ids: validImages.map( ( { id } ) => id ),
-						} );
-					}
-					return createBlock( `coblocks/${ name }` );
-				},
-			},
-			{
-				type: 'prefix',
-				prefix: ':masonry',
-				transform: function( content ) {
-					return createBlock( `coblocks/${ name }`, {
-						content,
-					} );
-				},
-			},
-		],
-		to: [
-			{
-				type: 'block',
-				blocks: [ 'core/gallery' ],
-				transform: ( attributes ) => (
-					createBlock( 'core/gallery', {
-						...GalleryTransforms( attributes ),
-					} )
-				),
-			},
-		],
-	},
+	transforms,
 
 	edit,
 
 	save( { attributes, className } ) {
 
 		const {
+			captions,
 			gridSize,
 			gutter,
 			gutterMobile,
 			images,
 			linkTo,
-			captions,
 		} = attributes;
 
 		const innerClasses = classnames(
